+++ pytsite/auth/__init__.py
@@ -6,7 +6,8 @@ from ._api import get_current_user, get_user_statuses, get_user, create_user, ge
     user_nickname_rule, sign_in, get_auth_driver, create_role, verify_password, hash_password, sign_out, \
     get_access_token_info, switch_user, get_anonymous_user, get_system_user, get_users, get_storage_driver, \
     count_users, count_roles, get_first_admin_user, get_roles, switch_user_to_system, switch_user_to_anonymous, \
-    restore_user, generate_access_token, prolong_access_token, register_storage_driver, get_auth_drivers
+    restore_user, generate_access_token, prolong_access_token, register_storage_driver, get_auth_drivers, \
+    revoke_access_token
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'