+++ pytsite/auth_storage_odm/_field.py
@@ -60,7 +60,7 @@ class Roles(_odm.field.UniqueStringList):
 
 
 class User(_odm.field.Abstract):
-    """Field to store reference to user.
+    """Field to store reference to user
     """
 
     def __init__(self, name: str, **kwargs):
@@ -73,34 +73,51 @@ class User(_odm.field.Abstract):
         """
         self._allow_anonymous = kwargs.get('allow_anonymous', False)
         self._allow_system = kwargs.get('allow_system', False)
+        self._disallowed_users = kwargs.get('disallowed_users', ())
 
         super().__init__(name, **kwargs)
 
-    def _resolve_user_uid(self, value) -> str:
-        if isinstance(value, str):
+    def _resolve_user(self, value) -> _auth.model.AbstractUser:
+        if isinstance(value, _auth.model.AbstractUser):
             return value
+        elif isinstance(value, str):
+            return _auth.get_user(uid=value)
         elif isinstance(value, _DBRef):
-            return str(value.id)
-        elif isinstance(value, _auth.model.AbstractUser):
-            if value.is_anonymous:
-                if not self._allow_anonymous:
-                    raise ValueError('Anonymous user is not allowed here')
-                return 'ANONYMOUS'
-            elif value.is_system:
-                if not self._allow_system:
-                    raise ValueError('System user is not allowed here')
-                return 'SYSTEM'
-            else:
-                return value.uid
+            return _auth.get_user(uid=value.id)
         else:
             raise TypeError("Field '{}': user object, str or DB ref expected, got {}.".
                             format(self._name, type(value)))
 
+    def _check_user(self, user: _auth.model.AbstractUser):
+        if user.is_anonymous:
+            if not self._allow_anonymous:
+                raise ValueError('Anonymous user is not allowed here')
+            return 'ANONYMOUS'
+        elif user.is_system:
+            if not self._allow_system:
+                raise ValueError('System user is not allowed here')
+            return 'SYSTEM'
+
+        if self._disallowed_users:
+            for user in self._disallowed_users:  # type: _auth.model.AbstractUser
+                if user.uid == user.uid:
+                    raise ValueError("User '{}' is not allowed here".format(user.login))
+
+        return user
+
     def _on_set(self, value, **kwargs) -> str:
         """Hook
+
+        Internally this field stores only user's UID as string.
         """
-        # Internally this field stores only user's UID as string
-        return self._resolve_user_uid(value)
+        value = self._check_user(self._resolve_user(value))
+
+        if value.is_anonymous:
+            return 'ANONYMOUS'
+        elif value.is_system:
+            return 'SYSTEM'
+
+        return value.uid
 
     def _on_get(self, value: str, **kwargs) -> _Optional[_auth.model.AbstractUser]:
         """Hook
@@ -149,9 +166,9 @@ class Users(User):
         """Hook
         """
         if not isinstance(value, (list, tuple)):
-            raise TypeError("Field '{}': list or tuple expected, got {}.".format(self.name, type(value)))
+            raise TypeError("Field '{}': list or tuple expected, got {}".format(self.name, type(value)))
 
-        return [self._resolve_user_uid(v) for v in value]
+        return [self._check_user(self._resolve_user(v)).uid for v in value]
 
     def _on_get(self, value: list, **kwargs) -> _Tuple[_auth.model.AbstractUser, ...]:
         """Hook
@@ -166,11 +183,11 @@ class Users(User):
         return tuple(r)
 
     def _on_add(self, current_value: list, raw_value_to_add, **kwargs):
-        current_value.append(self._resolve_user_uid(raw_value_to_add))
+        current_value.append(self._check_user(self._resolve_user(raw_value_to_add)).uid)
 
         return current_value
 
     def _on_sub(self, current_value: list, raw_value_to_sub, **kwargs):
-        raw_value_to_sub = self._resolve_user_uid(raw_value_to_sub)
+        raw_value_to_sub = self._check_user(self._resolve_user(raw_value_to_sub)).uid
 
         return [uid for uid in current_value if uid != raw_value_to_sub]